card-builder.tsの作成
表示用にデータを触る関数群
